Implement two gizmos for the 3D editor view, one to zoom and one to pan
This implementation features the addition of 2 new gizmos on the 3d
editor view on the top right. These 2 gizmos that can be dragged on
either with a mouse or touch to either zoom in or out (for the first
gizmo) or pan the camera vertically or horizontally (for the second
gizmo) depending on the direction of the drag. These can be enabled and
disabled in the Editor Settings in the tab Editors/3D/Navigation.

Co-authored-by:	Guilherme Luis Rogão Viana <guilherme.viana@tecnico.ulisboa.pt>